Instead of evaluating JS, get JSON instead
This removes the concept of processing from the works, as this was dead
code anyway.
