+++ app/views/users/_notify_number.html.erb
@@ -1,3 +1,3 @@
-<span id="notify_number" class="label label-<%= (@notify_number || 0) > 0 ? 'danger' : 'default' %>" data-update-poll-url="<%= sufia.user_notify_path %>" data-update-poll-interval="30">
-  <%= @notify_number %> <span class="sr-only">unread notifications</span>
+<span id="notify_number" class="label label-default" data-update-poll-url="<%= sufia.user_notify_path %>" data-update-poll-interval="30">
+  <span class="count">0</span> <span class="sr-only">unread notifications</span>
 </span>
