--- app/views/users/notifications_number.js.erb
@@ -1,6 +1,0 @@
-$("#notify_number").replaceWith("<%= escape_javascript(render partial: "notify_number") %>");
-<%# remove the icon that indicates the upload sets is waiting for background
-    processing once the process has completed %>
-<% @processed_works.each do |upload_set|%>
-  $(".<%= upload_set %>").attr('class', '');
-<% end %>
