+++ app/views/users/notifications_number.json.jbuilder
@@ -1,0 +1,1 @@
+json.notify_number @notify_number
