+++ spec/factories/users.rb
@@ -5,15 +5,11 @@ FactoryGirl.define do
 
     factory :user_with_mail do
       after(:create) do |user|
-        # TODO: what is this class for?
-        # <span class="batchid ui-helper-hidden">fake_upload_set_id</span>
-        message = UploadSetMessage.new
-
         # Create examples of single file successes and failures
         (1..10).each do |number|
           file = MockFile.new(number.to_s, "Single File #{number}")
-          User.batchuser.send_message(user, message.single_success("single-batch-success", file), message.success_subject, false)
-          User.batchuser.send_message(user, message.single_failure("single-batch-failure", file), message.failure_subject, false)
+          User.batchuser.send_message(user, 'File 1 could not be updated. You do not have sufficient privileges to edit it.', file.to_s, false)
+          User.batchuser.send_message(user, 'File 1 has been saved', file.to_s, false)
         end
 
         # Create examples of mulitple file successes and failures
@@ -21,8 +17,8 @@ FactoryGirl.define do
         (1..50).each do |number|
           files << MockFile.new(number.to_s, "File #{number}")
         end
-        User.batchuser.send_message(user, message.multiple_success("multiple-batch-success", files), message.success_subject, false)
-        User.batchuser.send_message(user, message.multiple_failure("multiple-batch-failure", files), message.failure_subject, false)
+        User.batchuser.send_message(user, 'These files could not be updated. You do not have sufficient privileges to edit them.', 'Batch upload permission denied', false)
+        User.batchuser.send_message(user, 'These files have been saved', 'Batch upload complete', false)
       end
     end
   end
@@ -34,8 +30,4 @@ class MockFile
     self.id = id
     self.to_s = string
   end
-end
-
-class UploadSetMessage
-  include CurationConcerns::Messages
 end
